+++ OKX.Api/Financial/Responses/OkxFinancialFlexibleLoanCollateralAssets.cs
@@ -1,0 +1,13 @@
+namespace OKX.Api.Financial;
+
+/// <summary>
+/// Flexible loan collateral assets response
+/// </summary>
+public record OkxFinancialFlexibleLoanCollateralAssets
+{
+    /// <summary>
+    /// Available collateral assets
+    /// </summary>
+    [JsonProperty("assets")]
+    public List<OkxFinancialFlexibleLoanAsset> Assets { get; set; } = [];
+}